Factor out command line template implementations.
