+++ src/utils/ArgsParse.cpp
@@ -268,139 +268,11 @@ int ArgsParser::OptionalArg::compareWithRequired(const RequiredArg& R) const {
   return Arg::compare(R);
 }
 
-template <>
-bool ArgsParser::Optional<bool>::select(charstring OptionValue) {
-  Value = !DefaultValue;
-  return false;
-}
-
-template <>
-void ArgsParser::Optional<bool>::describeDefault(FILE* Out,
-                                                 size_t TabSize,
-                                                 size_t& Indent) const {
-  printDescriptionContinue(Out, TabSize, Indent, " (default is ");
-  printDescriptionContinue(Out, TabSize, Indent,
-                           DefaultValue ? "true" : "false");
-  printDescriptionContinue(Out, TabSize, Indent, ")");
-}
-
-bool ArgsParser::Toggle::select(charstring OptionValue) {
-  Value = !Value;
-  return false;
-}
-
-void ArgsParser::Toggle::describeDefault(FILE* Out,
-                                         size_t TabSize,
-                                         size_t& Indent) const {
-  Optional<bool>::describeDefault(Out, TabSize, Indent);
-  printDescriptionContinue(Out, TabSize, Indent,
-                           " (each occurrence toggles value)");
-}
-
-template <>
-bool ArgsParser::Optional<charstring>::select(charstring OptionValue) {
-  Value = OptionValue;
-  return true;
-}
-
-template <>
-void ArgsParser::Optional<charstring>::describeDefault(FILE* Out,
-                                                       size_t TabSize,
-                                                       size_t& Indent) const {
-  if (DefaultValue == nullptr) {
-    printDescriptionContinue(Out, TabSize, Indent, " (has no default value)");
-    return;
-  }
-  printDescriptionContinue(Out, TabSize, Indent, " (default is '");
-  printDescriptionContinue(Out, TabSize, Indent, DefaultValue);
-  printDescriptionContinue(Out, TabSize, Indent, "')");
-}
-
-template <>
-bool ArgsParser::Optional<uint32_t>::select(charstring OptionValue) {
-  Value = size_t(atoll(OptionValue));
-  return true;
-}
-
-template <>
-void ArgsParser::Optional<uint32_t>::describeDefault(FILE* Out,
-                                                     size_t TabSize,
-                                                     size_t& Indent) const {
-  printDescriptionContinue(Out, TabSize, Indent, " (default is ");
-  writeSize_t(Out, TabSize, Indent, DefaultValue);
-  printDescriptionContinue(Out, TabSize, Indent, ")");
-}
-
-template <>
-bool ArgsParser::SetValue<uint32_t>::select(charstring OptionValue) {
-  Value = SelectValue;
-  return false;
-}
-
-template <>
-void ArgsParser::Optional<int32_t>::describeDefault(FILE* Out,
-                                                    size_t TabSize,
-                                                    size_t& Indent) const {
-  printDescriptionContinue(Out, TabSize, Indent, " (default is ");
-  writeSize_t(Out, TabSize, Indent, DefaultValue);
-  printDescriptionContinue(Out, TabSize, Indent, ")");
-}
-
-template <>
-bool ArgsParser::Optional<uint64_t>::select(charstring OptionValue) {
-  Value = size_t(atoll(OptionValue));
-  return true;
-}
-
-template <>
-void ArgsParser::Optional<uint64_t>::describeDefault(FILE* Out,
-                                                     size_t TabSize,
-                                                     size_t& Indent) const {
-  printDescriptionContinue(Out, TabSize, Indent, " (default is ");
-  writeSize_t(Out, TabSize, Indent, DefaultValue);
-  printDescriptionContinue(Out, TabSize, Indent, ")");
-}
-
-template <>
-bool ArgsParser::Optional<int64_t>::select(charstring OptionValue) {
-  Value = size_t(atoll(OptionValue));
-  return true;
-}
-
-template <>
-void ArgsParser::Optional<int64_t>::describeDefault(FILE* Out,
-                                                    size_t TabSize,
-                                                    size_t& Indent) const {
-  printDescriptionContinue(Out, TabSize, Indent, " (default is ");
-  writeSize_t(Out, TabSize, Indent, DefaultValue);
-  printDescriptionContinue(Out, TabSize, Indent, ")");
-}
-
 int ArgsParser::RequiredArg::compare(const Arg& A) const {
   if (isa<OptionalArg>(A))
     return -cast<OptionalArg>(&A)->compareWithRequired(*this);
 
   return Arg::compare(A);
-}
-
-template <>
-bool ArgsParser::Required<charstring>::select(charstring OptionValue) {
-  Value = OptionValue;
-  return true;
-}
-
-template <>
-bool ArgsParser::RepeatableSet<std::string>::select(charstring Add) {
-  Values.insert(Add);
-  return true;
-}
-
-template <>
-void ArgsParser::RepeatableSet<std::string>::describeDefault(
-    FILE* Out,
-    size_t TabSize,
-    size_t& Indent) const {
-  printDescriptionContinue(Out, TabSize, Indent, " (can be repeated)");
 }
 
 ArgsParser::ArgsParser(charstring Description)
